Updates buildkit, Sonico, gitignroe and config
Ships the config with builds, fixes wording of prints on bot, removes unneeded comments removes usesless new lines

buildkit now detects Python versions and the OS the builder is using, stops the buildkit from erroring while trying to start the bot. 
Adds 'rstrip' to the automatic token importer, this stops it from adding \n to the token, stopping the bot from logging in.

Gitignore nolonger blogs the config (it doesn't contain tokens, so it doesn't need to be hidden), details (which contains tokens for each builder) has been added to gitignore

Config.json has had 'v' removed from the version (The bot adds version to the string before showing it to the user, and has the new build information from me working on it.

This commits need to be synced with the dev branch, it has been pushed here because the issues are higher ranked than feature updates.